Start refactoring and code splitting
